requirements.txt
DateTime==5.5
requests==2.31.0
mysql-connector-python==8.3.0